Replace `root_project_or_failed_lazytest`
