--- test/pkgs/sample/PkgWithoutProject/src/PkgWithoutProject.jl
@@ -1,1 +1,0 @@
-module PkgWithoutProject end
